+++ CHANGELOG.md
@@ -14,9 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - auto buy upgrades
 - farm mode
 
-## [Unreleased] (done)
+## [0.2.0] - 2025-10-22
 ## Changed
 - don't "fetch" the whole ghostimage each time we place pixels
+## Added
+- Automatic relogging
 
 ## [0.1.0] - 2025-10-06
 ## Added